requirements.txt
altgraph==0.17.4
asgiref==3.7.2
cachetools==5.3.2
contourpy==1.2.0
cx-Freeze==6.15.12
cx-Logging==3.1.0
cycler==0.12.1
Django==4.2.6
fonttools==4.47.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
kiwisolver==1.4.5
lief==0.13.2
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pefile==2023.2.7
pillow==10.2.0
py2exe==0.13.0.1
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
typing_extensions==4.8.0
tzdata==2023.3
zipp==3.17.0
